refactor(agents): restructure mcp module (enablement, runtime, shared) (NV-8053)
Move enable/disable use cases to mcp/enablement and mechanical helpers to
mcp/shared so runtime services own Anthropic projection policy.
