+++ apps/api/src/app/agents/management/agent-runtime.controller.ts
@@ -17,7 +17,7 @@ import {
 } from '@nestjs/common';
 import { ApiExcludeController, ApiOperation } from '@nestjs/swagger';
 import { RequirePermissions } from '@novu/application-generic';
-import { ApiRateLimitCategoryEnum, ApiAuthSchemeEnum, PermissionsEnum, UserSessionData } from '@novu/shared';
+import { ApiAuthSchemeEnum, ApiRateLimitCategoryEnum, PermissionsEnum, UserSessionData } from '@novu/shared';
 import type { Request } from 'express';
 import { getClientIp } from 'request-ip';
 import { RequireAuthentication } from '../../auth/framework/auth.decorator';
@@ -36,16 +36,16 @@ import { EnsureProviderManagedVaultCommand } from '../mcp/connections/ensure-pro
 import { EnsureProviderManagedVault } from '../mcp/connections/ensure-provider-managed-vault/ensure-provider-managed-vault.usecase';
 import { GetMcpConnectionStatusCommand } from '../mcp/connections/get-mcp-connection-status/get-mcp-connection-status.command';
 import { GetMcpConnectionStatus } from '../mcp/connections/get-mcp-connection-status/get-mcp-connection-status.usecase';
+import { DisableAgentMcpServerCommand } from '../mcp/enablement/disable-agent-mcp-server/disable-agent-mcp-server.command';
+import { DisableAgentMcpServer } from '../mcp/enablement/disable-agent-mcp-server/disable-agent-mcp-server.usecase';
+import { EnableAgentMcpServerCommand } from '../mcp/enablement/enable-agent-mcp-server/enable-agent-mcp-server.command';
+import { EnableAgentMcpServer } from '../mcp/enablement/enable-agent-mcp-server/enable-agent-mcp-server.usecase';
+import { ListAgentMcpServersCommand } from '../mcp/enablement/list-agent-mcp-servers/list-agent-mcp-servers.command';
+import { ListAgentMcpServers } from '../mcp/enablement/list-agent-mcp-servers/list-agent-mcp-servers.usecase';
+import { SetAgentMcpServersCommand } from '../mcp/enablement/set-agent-mcp-servers/set-agent-mcp-servers.command';
+import { SetAgentMcpServers } from '../mcp/enablement/set-agent-mcp-servers/set-agent-mcp-servers.usecase';
 import { GenerateMcpOAuthUrlCommand } from '../mcp/oauth/generate-mcp-oauth-url/generate-mcp-oauth-url.command';
 import { GenerateMcpOAuthUrl } from '../mcp/oauth/generate-mcp-oauth-url/generate-mcp-oauth-url.usecase';
-import { DisableAgentMcpServerCommand } from '../mcp/servers/disable-agent-mcp-server/disable-agent-mcp-server.command';
-import { DisableAgentMcpServer } from '../mcp/servers/disable-agent-mcp-server/disable-agent-mcp-server.usecase';
-import { EnableAgentMcpServerCommand } from '../mcp/servers/enable-agent-mcp-server/enable-agent-mcp-server.command';
-import { EnableAgentMcpServer } from '../mcp/servers/enable-agent-mcp-server/enable-agent-mcp-server.usecase';
-import { ListAgentMcpServersCommand } from '../mcp/servers/list-agent-mcp-servers/list-agent-mcp-servers.command';
-import { ListAgentMcpServers } from '../mcp/servers/list-agent-mcp-servers/list-agent-mcp-servers.usecase';
-import { SetAgentMcpServersCommand } from '../mcp/servers/set-agent-mcp-servers/set-agent-mcp-servers.command';
-import { SetAgentMcpServers } from '../mcp/servers/set-agent-mcp-servers/set-agent-mcp-servers.usecase';
 import { AgentRuntimeExceptionFilter } from '../shared/agent-runtime-exception.filter';
 import {
   AgentMcpServerEnablementResponseDto,
