+++ apps/api/src/app/agents/mcp/enablement/disable-agent-mcp-server/disable-agent-mcp-server.command.ts
