+++ apps/api/src/app/agents/mcp/enablement/enable-agent-mcp-server/enable-agent-mcp-server.command.ts
