+++ apps/api/src/app/agents/mcp/enablement/set-agent-mcp-servers/set-agent-mcp-servers.command.ts
