+++ apps/api/src/app/agents/mcp/shared/assert-mcp-novu-app-flag-enabled.ts
