+++ apps/api/src/app/agents/mcp/shared/assert-mcp-provider-managed-flag-enabled.ts
