+++ apps/api/src/app/agents/mcp/shared/resolve-managed-agent-always-allow-tool-permissions.ts
