+++ apps/api/src/app/agents/mcp/shared/resolve-mcp-servers.ts
