+++ apps/api/src/app/agents/mcp/shared/sanitize-url-for-logging.spec.ts
